ELEC-37 Refactored x86 timers to be more efficient, cleaned up comments, unity_extra -> test_helpers, validated SYSCLK for timers, minor bugfixes.

+++ libraries/ms-common/src/pqueue_backed.c
@@ -19,7 +19,7 @@ StatusCode pqueue_backed_init_impl(PQueueBacked *queue, PQueueNode *nodes, void 
   return STATUS_CODE_OK;
 }
 
-StatusCode pqueue_backed_push(PQueueBacked *queue, void *elem, uint16_t prio) {
+StatusCode pqueue_backed_push(PQueueBacked *queue, void *elem, uint32_t prio) {
   if (elem == NULL) {
     return status_code(STATUS_CODE_INVALID_ARGS);
   }
